requirements.txt
streamlit>=1.41.0
langgraph>=0.2.60
langchain>=0.3.15
pandas>=2.2.0
numpy>=1.26.0
requests>=2.32.0
tqdm>=4.67.0
neo4j>=5.26.0
chromadb>=0.5.23
google-genai>=1.2.0
matplotlib>=3.9.0
seaborn>=0.13.2
scipy>=1.13.0
scikit-learn>=1.5.0
lifelines>=0.30.0
arize-phoenix>=4.30.0
openinference-instrumentation-langchain>=0.1.34
opentelemetry-sdk>=1.30.0
opentelemetry-exporter-otlp>=1.30.0
